Use correct lifecycle phase to create temp file
The temp file was being created in the initialization phase, when the CompileTypeScript task is constructed, but it was being designated for deletion during the execution phase. The execution phase may not happen if the task is determined to be up-to-date. This leaves empty compiler documents in the temp folder.

Might as well explicitly delete the file when we’re done with it. If we’re using Gradle daemon, it can take a long time before the JVM exits, and keeping a bunch of files open could lead to interesting failures.